requirements.txt
python-dotenv==1.0.1
langchain==0.3.7
langchain-huggingface==0.1.2
langchain-core==0.3.15
langchain-openai==0.2.3
langchain-ollama==0.2.0
langchain_groq==0.3.8
langchain-pinecone==0.2.0
langchain-community==0.3.2
langchain-chroma==0.1.4
chromadb==0.5.15
torch==2.2.2
sentence-transformers==3.2.1
llama-cpp-python==0.3.2
pypdf==4.3.1
docx2txt2==1.0.4
ragas==0.2.4
llama-index==0.11.22
nltk==3.9.1
bs4==0.0.2
tavily-python==0.5.0
google-search-results==2.4.2 #This Python package is meant to scrape and parse search results from Google, Bing, Baidu, Yandex, Yahoo, Home Depot, eBay and more, using SerpApi. https://serpapi.com
                             # Google Serper Dev is included in langchain_community.utilities GoogleSerperAPIWrapper module https://serper.dev
requests==2.32.3
wikipedia==1.4.0
scikit-learn~=1.5.2

# https://www.python.org/downloads/release/python-3127/
# https://www.python.org/downloads/release/python-3119/